update-python-libraries.py: Always use ref/tags/ for specifying rev
Avoids situations where a branch and tag share the same
label. Github will silently return a non-valid artifact and
will fail the build.

Since all github releases correspond to an annotated tag, it's
safe to assume that if we found a release for a tag, that the
tag will exist.